LayoutTests/svg/dom/SVGPoint-expected.txt

This test checks the SVGPoint API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial point values
PASS point.x is 0
PASS point.y is 0

Check assigning points
PASS point.x = 100 is 100
PASS point.y = 200 is 200

Check assigning invalid points
PASS point.x = point threw exception TypeError: Type error.
PASS point.y = null is null

Check that the point is still containing the correct values
PASS point.x is 100
PASS point.y is 0
PASS successfullyParsed is true

TEST COMPLETE
